Ran some unit tests on the contiguous float property, minor changes as a result.
This commit changes the toString function from a const char *
to a std::string to avoid pointing to stuff on the stack that
has been deallocated. Also ran clang format on the contiguous
float property without thinking, so that's why a bunch of formatting
stuff is changed. Finally added a getData function for the contiguous
float proprety that I will eventually replace with template magic in
the base property class.