Add a kill mode to the sheduler
The kill mode utilizes a "KillModeCommand" that when added to the
scheduler forces the scheduler into a special state. This state removes
all commands from the scheduler and calls all subsystems' "onEnterKillMode"
function. When in kill mode, the "onRefreshKillMode" function is called
repeatedly, and when the kill mode command is removed from the
scheduler, all subsystems' "onExitKillMode" function.