Use TCP for getting remote input
Added methods and fields to TCPServer class and added calls to TCPServer
methods in Remote.cpp so that remote now checks for TCPServer to
see if a message is ready then copies it. Next thing to do is add
a update() function for TCPServer to be called in main while loop
so that TCPServer can poll socket for new data.