Read remote input through TCP
Add functions to support reading input from TCP socket for remote
data. This will also allow adding functions for reading input at any
time through this new updateInput() function that gets called in the
main while loop.

I have some slight concerns as to how this is handled and do think
that multithreading would be the optimal way to go about this. But
looking at the default CAN-handling code a single/up-to-two CAN messages
are handled once per loop anyways so this shouldn't be too bad.

Next steps: I need to set up some test code for this, and then after
that it's probably a matter of getting the actual external remote
simulator (on Windows) to actually send messages and in the right
format.