Add TCPServer read unit test
Update a lot of functions to use the const keyword correctly
Struggling with threads, for some reason passing const char*
(a string literal) to a thread expecting a char* was giving this
"not an rvalue" error or something, but changing the function to
expect a const char* fixed it? I'm very confused on why, looking
at cppreference on what counts as an rvalue, string literal should
have worked? idk why changing the expected type fixed the error but it
did. so whatever.