Use threads instead of fork in TCP Unit test
Previously the "SendingCorrectMessages" test for the TCPServer used
fork to be able to get past a point in the logical execution between
client and server where they both would need to block for each other.
The issue with this however was that fork would create two instances
of the TCPServer and I didn't like that. Changed to use threads instead.