Fix "The command line is too long" error when linking on Windows
libc on Windows, or some related component, only accepts exec'd commands
up to around 8100 characters long. When scons invokes g++ to link a
final test executable, it includes paths for all object files as
parameters, which can make the executed command very long. To avoid the
error, we now take the advice of the SCons wiki and replace SCons'
built-in "spawn" function with a custom shim that invokes Popen and thus
is internally using the Windows API directly rather than libc.