Draft working 2022 sentinel code and constants
We had a crappy DR16 that caused mucho anguish. That sucked :/

Also a lot of friction on the yaw pulley for turret 1. Other big
ME issue to resolve is feeding tubes have a lot of internal
resistance to passing balls. Turret 2 jams so bad it's hard to
unjam even when shoving a screwdriver in there.