Flip signs on x and y to make hero drive correctly
Not sure if this is affected by the chassis refactor stuff. Basically
the hero would drive backward when trying to go forward, left when
right, etc. By swapping the sign for the x and y values in the function,
i was able to get it to drive correctly. Also the
wheelVelToChassisVelMat set in MechaniumChassisSubsystem doesnt seem to
be used when driving with a remote, which seems wrong.